% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.ladie.R
\name{AIC.ladie}
\alias{AIC.ladie}
\title{Akaike Information Criterion for class ladie}
\usage{
\method{AIC}{ladie}(object)
}
\arguments{
\item{object}{object of class ladie}
}
\description{
AIC is given by \eqn{-2\mbox{loglik}(\hat\theta) + 2 * |\theta|}
}
