% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.ladie.R
\name{BIC.ladie}
\alias{BIC.ladie}
\title{Bayesian Information Criterion for class ladie}
\usage{
\method{BIC}{ladie}(object)
}
\arguments{
\item{object}{object of class ladie}
}
\description{
BIC is given by \eqn{-2\mbox{loglik}(\hat\theta) + \log(n) * |\theta|}
}
