% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_SUBSET_multipath_matrix.R
\name{create_SUBSET_multipath_matrix}
\alias{create_SUBSET_multipath_matrix}
\title{Helper function for SUBSET_multipath}
\usage{
create_SUBSET_multipath_matrix(objects)
}
\description{
This function is designed to help individuals create a matrix for the
argument paths_by_variable in SUBSET_multipath via prompts.
}
