% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HQC.ladie.R
\name{HQC.ladie}
\alias{HQC.ladie}
\title{Hannon Quinn Information Criterion for class ladie}
\usage{
\method{HQC}{ladie}(object)
}
\arguments{
\item{object}{object of class ladie}
}
\description{
HQC is given by \eqn{-2\mbox{loglik}(\hat\theta) + 2 * |\theta|}
}
