% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladie_boot.R
\name{ladie_boot}
\alias{ladie_boot}
\title{ladie: LAtent Dose Incidence Estimation}
\usage{
ladie_boot(
  formula,
  data,
  dose_response = c("beta-poisson", "exponential")[1],
  prior_regr_coefs = list(mean = 0, sd = 2.5, autoscale = TRUE),
  prior_regr_intercept = list(mean = 0, sd = 10),
  prior_sigma_df = 3,
  prior_alpha_rate = 1,
  nonlinear_time = FALSE,
  CI_level = 0.95,
  n_bootstraps = 250,
  verbose = TRUE,
  cluster
)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted.}

\item{data}{a data.frame containing the variables in the model.}

\item{dose_response}{The dose-response model to be used.  Either the
beta-poisson or exponential.}

\item{prior_regr_coefs}{a named list giving the hyperparameters of the
regression coefficients, excluding the intercept.  Should include the following:
\itemize{
\item mean - mean of the normal prior
\item sd - standard deviation of the norma prior
\item autoscale - logical used to determine whether the prior sd should be
rescaled according to the standard deviation of the columns of the design matrix
}}

\item{prior_regr_intercept}{named list, same as prior_regr_coefs, except no autoscale}

\item{prior_sigma_df}{positive number giving the degrees of freedom for the half-t
prior on the standard deviation of the mean dose distribution}

\item{prior_alpha_rate}{positive value giving the rate of the exponential
prior on alpha (for beta-poisson model only)}

\item{nonlinear_time}{logical.  Set to TRUE if the expected dose should be the rate
times time raised to some power to be estimated. NOTE: this loses the interpretability
of the regression coefficients somewhat.}

\item{CI_level}{level for the credible intervals.}

\item{n_bootstraps}{integer. Number of Bayesian bootstrap samples}

\item{verbose}{logical. Should progress be displayed?}

\item{cluster}{integer giving the number of threads to use, or more directly an
object of class "SOCKcluster" from parallel::makeCluster().}
}
\value{
An object of class "ladie", which is really a list with the following
named elements:
\itemize{
\item summary - data.frame giving the posterior median, the posterior SD,
the credible intervals (at level CI_level), and the probability of direction.
\item CI_level
\item log_posterior - value of the log posterior evaluated at the MAP.
Useful for information criteria.
\item asymptotic_covariance - Full posterior covariance matrix.
}
}
\description{
Use longitudinal data to estimate risk ratios of accruing dose.  ladie_boot uses
Bayesian bootstrapping to perform estimation.
}
