% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUBSET_multipath.R
\name{SUBSET_multipath}
\alias{SUBSET_multipath}
\title{SUBSET for multiple pathogens}
\usage{
SUBSET_multipath(objects, paths_by_variable, n_prior_draws = 500, nu_max, nu)
}
\arguments{
\item{objects}{list of objects of class ladie}

\item{paths_by_variable}{binary 0/1 matrix.  The rows should correspond 1-to-1 with the
covariates (including intercept) for each model.  The columns should correspond
1-to-1 with the pathogens represented by each of the objects.  If two or more
entries in a row are 1, then these coefficients from the corresponding pathogens
will be shrunk towards each other.  All 0 entries do not undergo any shrinkage estimates
for that covariate.}

\item{nu_max}{positive numeric giving the maximum level of shrinkage allowed.}

\item{nu}{positive value giving the exact level of shrinkage.}
}
\value{
updated list of objects of class ladie
}
\description{
Implement the SUBSET approach (SUBspace Shrinkage via Exponential Tilting)
towards setting the regression coefficients (sans intercepts) obtained
from multiple pathogens proportional to each other.
}
