% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ladie.R
\name{summary.ladie}
\alias{summary.ladie}
\title{Summarizing LADIE model fits}
\usage{
summary.ladie(object, CI_level, print = TRUE)
}
\arguments{
\item{object}{object of class "ladie"}

\item{CI_level}{the level of credible intervals to be provided}

\item{print}{logical}
}
\description{
summary method for class "ladie"
}
